fix: 补全 God Object 拆分文件 + 修复 9 个 stale 测试引用
- 恢复 folder_scan_service 拆分（_candidate_post_processor + _import_pipeline）
- 恢复 document_template_admin 拆分（views_mixin + display_mixin）
- 修复 5 个测试文件的旧目录引用（repos/usecases/assemblers/admin/workflows → 新路径）
