+++ backend/apps/document_parsing/api/parsing_api.py
@@ -5,6 +5,8 @@ from pathlib import Path
 from typing import Optional
 
 from django.conf import settings
+from django.core.files.base import ContentFile
+from django.core.files.storage import default_storage
 from ninja import File, Router, UploadedFile
 
 from apps.core.security.auth import JWTOrSessionAuth
@@ -34,14 +36,9 @@ def parse_document(request: object, file: UploadedFile = File(...), body: ParseD
     """
     try:
         # 保存上传的文件
-        upload_dir = Path(settings.MEDIA_ROOT) / "document_parsing" / "uploads"
-        upload_dir.mkdir(parents=True, exist_ok=True)
-
         file_name = file.name or "uploaded"
-        file_path = upload_dir / file_name
-        with open(file_path, "wb") as f:
-            for chunk in file.chunks():
-                f.write(chunk)
+        saved_name = default_storage.save(f"document_parsing/uploads/{file_name}", file)
+        file_path = Path(settings.MEDIA_ROOT) / saved_name
 
         # 获取解析参数
         backend = body.backend if body else "auto"
@@ -85,14 +82,9 @@ def extract_text(request: object, file: UploadedFile = File(...), body: ExtractT
     """提取文档的纯文本内容"""
     try:
         # 保存上传的文件
-        upload_dir = Path(settings.MEDIA_ROOT) / "document_parsing" / "uploads"
-        upload_dir.mkdir(parents=True, exist_ok=True)
-
         file_name = file.name or "uploaded"
-        file_path = upload_dir / file_name
-        with open(file_path, "wb") as f:
-            for chunk in file.chunks():
-                f.write(chunk)
+        saved_name = default_storage.save(f"document_parsing/uploads/{file_name}", file)
+        file_path = Path(settings.MEDIA_ROOT) / saved_name
 
         # 获取参数
         backend = body.backend if body else "auto"
